fix(recharge): 允许充值金额为浮点数；修复充值时订单为空仍插入的问题

+++ backend/base/src/application/commands/transaction.rs
@@ -1,9 +1,9 @@
 use uuid::Uuid;
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct RechargeCommand {
     pub session_id: String,
-    pub amount: u32,
+    pub amount: f64,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
